discount price amount convert to double

+++ app/Repositories/DiscountRepository.php
@@ -14,7 +14,7 @@ class DiscountRepository
         if ($discount == null) {
             //initially discount set to zero
             $partner_service['discount_price'] = 0;
-            $partner_service['discounted_price'] = $partner_service->prices;
+            $partner_service['discounted_price'] = (double)$partner_service->prices;
         } /**
          * partner service has discount
          */
@@ -23,8 +23,8 @@ class DiscountRepository
              * discount is in percentage
              */
             if ($discount->is_amount_percentage) {
-                $amount = ((float)$partner_service->prices * $discount->amount) / 100;
-                if ($discount->cap != 0 && $amount > (float)$discount->cap) {
+                $amount = ((double)$partner_service->prices * $discount->amount) / 100;
+                if ($discount->cap != 0 && $amount > (double)$discount->cap) {
                     $amount = $discount->cap;
                 }
                 $partner_service['cap'] = $discount->cap;
@@ -32,8 +32,8 @@ class DiscountRepository
                 $partner_service['discounted_price'] = $partner_service->prices - $amount;
             } else {
                 $partner_service['cap'] = null;
-                $partner_service['discount_price'] = $discount->amount;
-                $partner_service['discounted_price'] = $partner_service->prices - $discount->amount;
+                $partner_service['discount_price'] = (double)$discount->amount;
+                $partner_service['discounted_price'] = (double)$partner_service->prices - $discount->amount;
                 $partner_service['discount_id'] = $discount->id;
             }
             if ($partner_service['discounted_price'] < 0) {
@@ -49,14 +49,14 @@ class DiscountRepository
      * @param $discount
      * @param $partnerPrice
      * @param $quantity
-     * @return float
+     * @return double
      * @internal param $hasPercentage
      * @internal param $discountValue
      */
     public function getDiscountAmount($discount, $partnerPrice, $quantity)
     {
         if ($discount['is_percentage']) {
-            $amount = ((float)$partnerPrice * $quantity) * ($discount['voucher']['amount'] / 100);
+            $amount = ((double)$partnerPrice * $quantity) * ($discount['voucher']['amount'] / 100);
             if ($discount['voucher']->cap != 0 && $amount > $discount['voucher']->cap) {
                 $amount = $discount['voucher']->cap;
             }
@@ -74,7 +74,7 @@ class DiscountRepository
     public function getServiceDiscountAmount($discount, $partnerPrice, $quantity)
     {
         if ($discount->is_amount_percentage) {
-            $amount = ((float)$partnerPrice * $quantity * $discount->amount) / 100;
+            $amount = ((double)$partnerPrice * $quantity * $discount->amount) / 100;
             if ($discount->cap != 0 && $amount > $discount->cap) {
                 $amount = $discount->cap;
             }
